split up comparators; improve key shape handling
