+++ spec/katachi_spec.rb
@@ -123,28 +123,28 @@ RSpec.describe Katachi do
     shape = { a: { b: [Kt.any_of(Integer, String)] } }
     expect(Kt.compare(value:, shape:).to_s).to eq <<~RESULT.chomp
       :hash_is_match <-- compare(value: {a: {b: [1, "a"]}}, shape: {a: {b: [AnyOf[Integer, String]]}})
-        :hash_has_no_missing_keys <-- compare(value: {a: {b: [1, "a"]}}, shape: {a: {b: [AnyOf[Integer, String]]}})
-          :hash_key_present <-- compare(value: :a, shape: :a)
-        :hash_has_no_extra_keys <-- compare(value: {a: {b: [1, "a"]}}, shape: {a: {b: [AnyOf[Integer, String]]}})
-          :hash_key_allowed <-- compare(value: :a, shape: :a)
-        :hash_values_are_match <-- compare(value: {a: {b: [1, "a"]}}, shape: {a: {b: [AnyOf[Integer, String]]}})
-          :kv_specific_match <-- compare(value: {a: {b: [1, "a"]}}, shape: {a: {b: [AnyOf[Integer, String]]}})
-            :hash_is_match <-- compare(value: {b: [1, "a"]}, shape: {b: [AnyOf[Integer, String]]})
-              :hash_has_no_missing_keys <-- compare(value: {b: [1, "a"]}, shape: {b: [AnyOf[Integer, String]]})
-                :hash_key_present <-- compare(value: :b, shape: :b)
-              :hash_has_no_extra_keys <-- compare(value: {b: [1, "a"]}, shape: {b: [AnyOf[Integer, String]]})
-                :hash_key_allowed <-- compare(value: :b, shape: :b)
-              :hash_values_are_match <-- compare(value: {b: [1, "a"]}, shape: {b: [AnyOf[Integer, String]]})
-                :kv_specific_match <-- compare(value: {b: [1, "a"]}, shape: {b: [AnyOf[Integer, String]]})
-                  :array_is_match <-- compare(value: [1, "a"], shape: [AnyOf[Integer, String]])
-                    :array_element_match <-- compare(value: 1, shape: [AnyOf[Integer, String]])
-                      :any_of_match <-- compare(value: 1, shape: [Integer, String])
-                        :match <-- compare(value: 1, shape: Integer)
-                        :mismatch <-- compare(value: 1, shape: String)
-                    :array_element_match <-- compare(value: "a", shape: [AnyOf[Integer, String]])
-                      :any_of_match <-- compare(value: "a", shape: [Integer, String])
-                        :mismatch <-- compare(value: "a", shape: Integer)
-                        :match <-- compare(value: "a", shape: String)
+        :hash_has_no_missing_keys <-- compare(value: {a: {b: [1, "a"]}}, shape: {a: {b: [AnyOf[Integer, String]]}}); child_label: :$required_keys
+          :hash_key_exact_match <-- compare(value: :a, shape: :a); child_label: :a
+        :hash_has_no_extra_keys <-- compare(value: {a: {b: [1, "a"]}}, shape: {a: {b: [AnyOf[Integer, String]]}}); child_label: :$extra_keys
+          :hash_key_exactly_allowed <-- compare(value: :a, shape: :a); child_label: :a
+        :hash_values_are_match <-- compare(value: {a: {b: [1, "a"]}}, shape: {a: {b: [AnyOf[Integer, String]]}}); child_label: :$values
+          :kv_specific_match <-- compare(value: {a: {b: [1, "a"]}}, shape: {a: {b: [AnyOf[Integer, String]]}}); child_label: {a: {b: [1, "a"]}}
+            :hash_is_match <-- compare(value: {b: [1, "a"]}, shape: {b: [AnyOf[Integer, String]]}); child_label: {b: [AnyOf[Integer, String]]}
+              :hash_has_no_missing_keys <-- compare(value: {b: [1, "a"]}, shape: {b: [AnyOf[Integer, String]]}); child_label: :$required_keys
+                :hash_key_exact_match <-- compare(value: :b, shape: :b); child_label: :b
+              :hash_has_no_extra_keys <-- compare(value: {b: [1, "a"]}, shape: {b: [AnyOf[Integer, String]]}); child_label: :$extra_keys
+                :hash_key_exactly_allowed <-- compare(value: :b, shape: :b); child_label: :b
+              :hash_values_are_match <-- compare(value: {b: [1, "a"]}, shape: {b: [AnyOf[Integer, String]]}); child_label: :$values
+                :kv_specific_match <-- compare(value: {b: [1, "a"]}, shape: {b: [AnyOf[Integer, String]]}); child_label: {b: [1, "a"]}
+                  :array_is_match <-- compare(value: [1, "a"], shape: [AnyOf[Integer, String]]); child_label: [AnyOf[Integer, String]]
+                    :array_element_match <-- compare(value: 1, shape: [AnyOf[Integer, String]]); child_label: 1
+                      :any_of_match <-- compare(value: 1, shape: [Integer, String]); child_label: AnyOf[Integer, String]
+                        :match <-- compare(value: 1, shape: Integer); child_label: Integer
+                        :mismatch <-- compare(value: 1, shape: String); child_label: String
+                    :array_element_match <-- compare(value: "a", shape: [AnyOf[Integer, String]]); child_label: "a"
+                      :any_of_match <-- compare(value: "a", shape: [Integer, String]); child_label: AnyOf[Integer, String]
+                        :mismatch <-- compare(value: "a", shape: Integer); child_label: Integer
+                        :match <-- compare(value: "a", shape: String); child_label: String
     RESULT
   end
